refactor: separate HIR introspection from VM formatting
Keep compiler-side inspection independent of the legacy backend so the symbolic WebAssembly target can format HIR without linking VM introspection.
